Create Api

+++ app/Http/Controllers/UserController.php
@@ -4,8 +4,17 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller{
+
+    public function __construct()
+    {
+        $this->middleware(function($request, $next){
+        if(Gate::allows('manage-users')) return $next($request);
+            abort(403, 'Anda tidak memiliki cukup hak akses');
+        });
+    }
     /**
      * Display a listing of the resource.
      *
@@ -46,6 +55,18 @@ class UserController extends Controller{
      */
     public function store(Request $request)
     {
+        $validation = \Validator::make($request->all(),[
+            "name" => "required|min:5|max:100",
+            "username" => "required|min:5|max:20|unique:users",
+            "roles" => "required",
+            "phone" => "required|digits_between:10,12",
+            "address" => "required|min:20|max:200",
+            "avatar" => "required",
+            "email" => "required|email|unique:users",
+            "password" => "required",
+            "password_confirmation" => "required|same:password"
+            ])->validate();
+
         $new_user = new \App\User;
         $new_user->name = $request->get('name');
         $new_user->username = $request->get('username');
@@ -97,21 +118,27 @@ class UserController extends Controller{
      */
     public function update(Request $request, $id)
     {
+        \Validator::make($request->all(), [
+            "name" => "required|min:5|max:100",
+            "roles" => "required",
+            "phone" => "required|digits_between:10,12",
+            "address" => "required|min:20|max:200",
+        ])->validate();
         $user = \App\User::findOrFail($id);
         $user->name = $request->get('name');
         $user->roles = json_encode($request->get('roles'));
         $user->address = $request->get('address');
         $user->phone = $request->get('phone');
-        if($user->avatar && file_exists(storage_path('app/public/' . $user->avatar))){
-        \Storage::delete('public/'.$user->avatar);
-        $file = $request->file('avatar')->store('avatars', 'public');
-        $user->avatar = $file;
+        $user->status = $request->get('status');
+        if($request->file('avatar')){
+            if($user->avatar && file_exists(storage_path('app/public/' . $user->avatar))){
+                \Storage::delete('public/'.$user->avatar);
+            }
+            $file = $request->file('avatar')->store('avatars', 'public');
+            $user->avatar = $file;
         }
         $user->save();
-        return redirect()->route('users.edit', ['id' => $id])->with('status','User succesfully updated');
-
-
-
+        return redirect()->route('users.edit', ['id' => $id])->with('status', 'User succesfully updated');
     }
 
     /**
